requirements.txt
aiosmtplib
pydantic-settings
fastapi
uvicorn
psycopg2-binary
SQLAlchemy>=2.0
alembic
passlib>=1
bcrypt==4.1.0
python-jose[cryptography]
pydantic
pydantic[email]
celery[redis]
python-dotenv
asyncpg>=0.27.0
